feat: supprimer un fichier ouvert renvoie vers la page 'Aucun fichier ouvert'
