+++ frontend/src/composables/Flowchart/useFlowChart.ts
@@ -6,14 +6,13 @@ import { useRoute, useRouter } from 'vue-router';
 import { computed, ref, watch } from 'vue';
 import { useShowErrorToast } from '../useShowErrorToast';
 import { FsEvent } from '@/types/FsEvent';
-import { node, watcher } from '$/models';
+import { watcher } from '$/models';
 import { Routes } from '@/types/Routes';
 
 export function useFlowChart() {
   const path = ref('');
   const route = useRoute();
   const router = useRouter()
-  const lastOp = ref<FsEvent | null>(null)
   const { updateNode, fromObject } = useVueFlow();
   const { showToast } = useShowErrorToast();
   useEventListener(window, 'paste', onPaste);
@@ -117,40 +116,8 @@ export function useFlowChart() {
     reader.readAsDataURL(file);
   }
 
-  function onFsEvent(e: FsEvent) {
-    lastOp.value = e
-    if (e.fileType != node.FileType.GRAPH) {
-      return
-    }
-
-    const routePath = route.params.path as string
-
-    let filePath = e.path + '/' + e.file
-    if (e.path === '.') {
-      filePath = e.file
-    }
-
-    if (
-      // If the operation is a delete but the deleted file is not the one currently opened, we skip
-      (filePath != routePath && e.op === watcher.Op.REMOVE) || // OR
-      // If the operation is a move or a rename and the old path is different from the one of the file currently opened, we skip
-      (e.oldPath != routePath.replace('./', '') &&
-        (e.op === watcher.Op.MOVE || e.op === watcher.Op.RENAME))
-    ) {
-      return;
-    }
-
-    if (e.op === watcher.Op.REMOVE) {
-      router.push({ name: Routes.NotOpened });
-      return;
-    }
-
-    router.push({ name: Routes.Flowchart, params: { path: filePath } });
-  }
-
   return {
     path,
     fileName,
-    onFsEvent
   };
 }
