+++ frontend/src/views/GraphPanel.vue
@@ -13,5 +13,39 @@
 
 <script setup lang="ts">
 import AppHeader from '@/components/AppHeader.vue';
-import { RouterView } from 'vue-router';
+import { RouterView, useRoute, useRouter } from 'vue-router';
+import { EventsOn } from '../../wailsjs/runtime/runtime';
+import { watcher } from '$/models';
+import { FsEvent } from '@/types/FsEvent';
+import { Routes } from '@/types/Routes';
+
+const route = useRoute();
+const router = useRouter();
+EventsOn('fsop', onFsEvent);
+
+function onFsEvent(e: FsEvent) {
+  const routePath = route.params.path as string;
+
+  let filePath = e.path + '/' + e.file;
+  if (e.path === '.') {
+    filePath = e.file;
+  }
+
+  if (
+    // If the operation is a delete but the deleted file is not the one currently opened, we skip
+    (filePath != routePath && e.op === watcher.Op.REMOVE) || // OR
+    // If the operation is a move or a rename and the old path is different from the one of the file currently opened, we skip
+    (e.oldPath != routePath.replace('./', '') &&
+      (e.op === watcher.Op.MOVE || e.op === watcher.Op.RENAME))
+  ) {
+    return;
+  }
+
+  if (e.op === watcher.Op.REMOVE) {
+    router.push({ name: Routes.NotOpened });
+    return;
+  }
+
+  router.push({ name: Routes.Flowchart, params: { path: filePath } });
+}
 </script>
